Fix test

+++ task1/task1_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestIsAutomorphic(t *testing.T) {
 	cases := []struct {
-		in   int
+		in   int64
 		want bool
 	}{
 		{1, true},
